+++ src/main/java/com/icomfortableworld/global/exception/GlobalExceptionHandler.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import com.icomfortableworld.common.dto.CommonResponseDto;
+import com.icomfortableworld.domain.member.exception.CustomMemberException;
 import com.icomfortableworld.global.exception.dto.ErrorResponseDto;
 import com.icomfortableworld.global.exception.feed.CustomFeedException;
 import com.icomfortableworld.global.exception.jwt.CustomJwtException;
-import com.icomfortableworld.global.exception.member.CustomMemberException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
